requirements.txt
asgiref==3.8.1
Django==5.1.8
django-cors-headers==4.7.0
django-filter==25.1
django-taggit==6.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
Markdown==3.7
packaging==24.2
pillow==11.1.0
PyJWT==2.9.0
sqlparse==0.5.3
